refactor(components): promote ui/primitives → atoms (phase 2)
Move the shared UI primitives (Button, IconButton, Input, Select, Card,
Toast/ToastViewport, cn) from components/ui/primitives to components/atoms,
establishing the atomic-design atoms layer. All 31 barrel importers updated
`@/components/ui/primitives` → `@/components/atoms` (no direct-path
importers existed). Pure move + import rename; no behavior change.

Verified via `tsc --noEmit`.
