+++ src/components/settings/SpaceMouseSettingsTab.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { ArrowLeftRight, Gamepad2, MousePointer2, SlidersHorizontal } from 'lucide-react';
 import type { SpaceMouseSettings } from '@/components/settings/spacemousePreferences';
-import { Select } from '@/components/ui/primitives';
+import { Select } from '@/components/atoms';
 
 type SpaceMouseSettingsTabProps = {
   settings: SpaceMouseSettings;
