+++ src/supports/Settings/SupportSidebar.tsx
@@ -39,7 +39,7 @@ import {
     GridSettingsCard,
     SupportKindTabs,
 } from './components';
-import { Card, CardHeader, IconButton } from '@/components/ui/primitives';
+import { Card, CardHeader, IconButton } from '@/components/atoms';
 import { NumberInput } from '@/components/ui/NumberInput';
 import { SelectDropdown } from '@/components/ui/SelectDropdown';
 import { SupportAnatomyPreviewSlot } from './AnatomyPreview/SupportAnatomyPreviewSlot';
